toolkit/content/tests/chrome/test_menu_hide.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css" type="text/css"?>

<window title="Menu Destruction Test"
  onload="setTimeout(runTests, 0);"
  xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">

  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>      
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"></script>      

<menu id="menu">
  <menupopup onpopupshown="this.firstChild.open = true" onpopuphidden="if (event.target == this) done()">
    <menu id="submenu" label="One">
      <menupopup onpopupshown="submenuOpened();">
        <menuitem label="Two"/>
      </menupopup>
    </menu>
  </menupopup>
</menu>

<script class="testbody" type="application/javascript">
<![CDATA[

SimpleTest.waitForExplicitFinish();

function runTests()
{
  $("menu").open = true;
}

function submenuOpened()
{
  var submenu = $("submenu")
  is(submenu.getAttribute('_moz-menuactive'), "true", "menu highlighted");
  submenu.hidden = true;
  $("menu").open = false;
}

function done()
{
  ok(!$("submenu").hasAttribute('_moz-menuactive'), "menu unhighlighted");
  SimpleTest.finish();
}

]]>
</script>

<body xmlns="http://www.w3.org/1999/xhtml">
<p id="display">
</p>
<div id="content" style="display: none">
</div>
<pre id="test">
</pre>
</body>

</window>
